Fix issue with gitlog when output contained utf8
When the git log command contained utf-8 characters, coveralls crashed with a
UnicodeDecodeError. This patch keeps the original behaviour unless the
UnicodeDecodeError is raised and will decode the text as utf-8 instead.

Fixing this also uncovered a UnicodeEncodeError, this has been fixed in the
same manner.

I also altered the test suite to use a commiter name that triggers this bug.

It might be cleaner to fix it by looking at the current locale or just assume
utf-8 but this seemed like the easiest and safest change for the time being.