test: remove test inter-dependencies
A bunch of tests had interdependencies since we rely on chdir'ing to
tempdirs (and then deleting those dirs) for various tests -- by running
the tests in randomized orders, we'd get randomized failures as it turns
out we were depending on eg. the existence of a previously-generated
.coverage file or being in a still-existing directory.

None of these randomized failures were actual functional issues, but
this should improve the consistency of our test runs in the future.